+++ src/appleseed/renderer/kernel/lighting/sppm/sppmlightingengine.cpp
@@ -445,7 +445,7 @@ namespace
 
                     // Evaluate the BSDF for this photon.
                     Spectrum bsdf_value;
-                    const double bsdf_prob =
+                    const float bsdf_prob =
                         vertex.m_bsdf->evaluate(
                             vertex.m_bsdf_data,
                             false,                                      // not adjoint
@@ -456,7 +456,7 @@ namespace
                             normalize(Vector3d(photon.m_incoming)),     // toward the light
                             ScatteringMode::Diffuse,
                             bsdf_value);
-                    if (bsdf_prob == 0.0)
+                    if (bsdf_prob == 0.0f)
                         continue;
 
                     // Make sure the BSDF value is spectral.
@@ -521,7 +521,7 @@ namespace
 
                     // Evaluate the BSDF for this photon.
                     Spectrum bsdf_value;
-                    const double bsdf_prob =
+                    const float bsdf_prob =
                         vertex.m_bsdf->evaluate(
                             vertex.m_bsdf_data,
                             false,                                      // not adjoint
@@ -532,7 +532,7 @@ namespace
                             normalize(Vector3d(photon.m_incoming)),     // toward the light
                             ScatteringMode::Diffuse,
                             bsdf_value);
-                    if (bsdf_prob == 0.0)
+                    if (bsdf_prob == 0.0f)
                         continue;
 
                     // The photons store flux but we are computing reflected radiance.
@@ -581,7 +581,7 @@ namespace
                 // Evaluate the environment EDF.
                 InputEvaluator input_evaluator(m_texture_cache);
                 Spectrum env_radiance;
-                double env_prob;
+                float env_prob;
                 m_env_edf->evaluate(
                     m_shading_context,
                     input_evaluator,
